services/api/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.6.1

SQLAlchemy==2.0.36
asyncpg==0.29.0

python-jose==3.3.0
passlib[bcrypt]==1.7.4

boto3==1.35.79

redis==5.0.8
rq==1.16.2

python-multipart==0.0.9
bcrypt==3.2.2